fix: hide delete checkbox on Retailer object change page
Apparently, we cannot set can_delete=False as an attribute in RetailerAdmin class like in django default inline admin classes.
As well as we cannot pass can_delete in inline_reverse as form and field passes. Which causes UserCreationForm fails to validate
when we tick the delete checkbox and hit save button in retailer admin page without entering new passwords in both password and
password confirmation fields (similar to when creating a new user and which can be inconvenient in this scenario).

So, I overridden the `ReverseInlineModelAdmin` class to set can_delete attribute to False to hide delete checkbox in Retailer object change page